pyproject.toml
[project]
name = "streamerdb"
version = "0.1"
dependencies = [
    "tortoise-orm"
]
requires-python = ">= 3.8"

[project.optional-dependencies]
testing = [
    "pytest",
    "pytest-asyncio",
]
